mysql에 맞게 member, pet column명 변경

+++ src/main/java/com/petchatbot/domain/dto/PetListDto.java
@@ -5,12 +5,12 @@ import lombok.Data;
 
 @Data
 public class PetListDto {
-    private Long petSerial;
+    private int petSerial;
     private Species petSpecies;
     private String petName;
 
 
-    public PetListDto(Long petSerial, Species petSpecies, String petName) {
+    public PetListDto(int petSerial, Species petSpecies, String petName) {
         this.petSerial = petSerial;
         this.petSpecies = petSpecies;
         this.petName = petName;
