+++ src/main/java/com/petchatbot/domain/model/Member.java
@@ -19,7 +19,7 @@ public class Member {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long member_serial;
+    private int member_serial;
     @Column(name = "member_email")
     private String memberEmail;
     @Column(name = "member_password")
